feat: add a new exec mode
Adds a new 'exec' mode that accepts two commands to control FPP.
The first command is the `data_cmd` and is used to provide the data that
FPP will operate on.
The second is the `exec_cmd` which will be called with the chosen files.

This also adds a new FPP option argument to control FPP's behavior. This
is useful when processing git status output which may have deleted
files, or for using it as a branch chooser.